requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
numpy==1.26.4
pillow==10.4.0
opencv-python-headless==4.10.0.84
scikit-image==0.24.0
mediapipe==0.10.14
